+++ src/states/GroundCling.ts
@@ -36,15 +36,13 @@ export default class GroundCling implements State {
 		{
 			this.jump();
 		}
-		else if (this.player.punchLeftInput == 'just-down' && this.player.punchCharged)
+		else if (this.player.punchInput == 'just-down' && this.player.punchCharged)
 		{
-			// this.player.flipX = false;
 			this.stateController.setState('punch');
 		}
-		else if (this.player.punchRightInput == 'just-down' && this.player.punchCharged)
+		else if (this.player.uppercutInput == 'just-down' && this.player.punchCharged)
 		{
-			// this.player.flipX = true;
-			this.stateController.setState('punch');
+			this.stateController.setState('uppercut');
 		}
 
 		if (!this.player.onFloor)